==> code/nbs/initial_workflow/0_preprocess_frames-echonet.ipynb <==
import os
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import cv2
import json
import glob

import torch
import torch.nn as nn
import torchvision
import torchvision.transforms as transforms
import torchvision.transforms.functional as TF

echonet_dir = '/zfs/wficai/Data/echonet_data/EchoNet-Dynamic/'
frame_export_dir = '/zfs/wficai/Data/echonet_data/frames/'
split_csv_export_dir = '/zfs/wficai/Data/echonet_data/'
os.chdir(echonet_dir)
d = pd.read_csv('FileList.csv')

res_x = 112
res_y = 112

d['vid_path'] = d.FileName.apply(lambda x: f"{echonet_dir}/Videos/{x}.avi")

d.head()

# standardize ESV and EDV
quantile_90 = pd.concat([d.ESV, d.EDV]).quantile(0.90)
quantile_90

d['ESV_scaled'] = d['ESV']/quantile_90
d['EDV_scaled'] = d['EDV']/quantile_90

for s in ('TRAIN', 'VAL', 'TEST'):
    d.query("Split==@s").to_csv(f"{split_csv_export_dir}/{s}.csv", index=None)

# Preprocess frames

def preproc(frames):

    frames = frames.permute(0,3,1,2)
    
    #Resize
    splitFrames = torch.split(frames, 10)
    resizedFrames = []
    aspectRatio = (frames.shape[3])/(frames.shape[2])
    for f in splitFrames:
        f = TF.resize(f, res_y)
        f = TF.center_crop(f, (res_x, res_y))
        
        resizedFrames.append(f)

    resizedFrames = torch.concat(resizedFrames)
    
    return resizedFrames

frame_list = []
for ix, row in d.iterrows():
    print(" "*100, end='\r')
    print(f"Converting video {ix+1} of {len(d)}.", end="\r")
    
    frames, _, _ =  torchvision.io.read_video(row['vid_path'], pts_unit = 'sec')
    frames = preproc(frames)
    
    # save individual frames
    for ix, f in enumerate(frames):
        filename = f"{frame_export_dir}{row['FileName']}_frame{str(ix).zfill(4)}.jpg"
        frame_list.append({'filename': row['FileName'], 'frame': ix, 'jpg_path': filename})
        torchvision.io.write_jpeg(f, filename)      

df_frames = pd.DataFrame(frame_list)
df_frames = df_frames.rename(columns={'filename': 'FileName'})
df_frames.to_csv(f"{split_csv_export_dir}/frames.csv", index=None)

df_frames
