==> code/nbs/initial_workflow/1_make_patient_study_table.ipynb <==
import pandas as pd
pd.set_option('display.max_rows', 200)

datestamp = '20221028'

df_pda = pd.read_csv('/zfs/wficai/pda/model_data/pda_study_patient_mapping.csv')
df_nopda = pd.read_csv('/zfs/wficai/pda/model_data/nonpda_study_patient_mapping.csv')

df_pda['patient_type'] = 'pda'
df_nopda['patient_type'] = 'nopda'
df = pd.concat([df_pda, df_nopda])

df['study'] = df['dicom_file_id'].apply(lambda x: x.split('_')[0])
df

df_patient_study = df.groupby(['patient_id', 'patient_type', 'study'], as_index=False).size().rename(columns={'size': 'num_videos'})
df_patient_study

df_patient_study.to_csv(f'/zfs/wficai/pda/model_data/{datestamp}_patient_study.csv', index=False)

df_patient = df_patient_study.groupby('patient_id', as_index=False).agg({'patient_type': lambda x: 'pda' if (x=='pda').any() else 'nopda', 'study': 'size'}).rename(columns = {'study': 'num_studies'})
df_patient.to_csv(f'/zfs/wficai/pda/model_data/{datestamp}_patient.csv', index=False)
df_patient

# number of studies by 'patient_id' -- 'patient_type'
# note: 'patient_type' should be called 'study_type' because pda status can (and does sometimes) change
df_patient_study.groupby(['patient_id', 'patient_type'], as_index=True)['study'].nunique()

# show this sorted to see if multiple studies tend to occur for pda or nonpda
# conclusion: mixed
df_patient_study.groupby(['patient_id', 'patient_type'], as_index=True)['study'].nunique().sort_values(ascending=False)

# overall number of studies by patient
df_patient_study.groupby(['patient_id'], as_index=True)['study'].nunique().sort_values(ascending=False)
