==> code/nbs/initial_workflow/4a_evaluate-multitask.ipynb <==
%load_ext autoreload
%autoreload 1
%config InlineBackend.print_figure_kwargs={'facecolor' : "w"}

import pandas as pd
import numpy as np
from torch.utils.data import DataLoader
import timm
from timm import optim, scheduler
import torch
from torchvision import transforms as tfm
from sklearn import metrics as skmet
import matplotlib.pyplot as plt
import json
import transforms as my_transforms

%aimport dataset
from models import MultiTaskFrameClassifier
ImageData = dataset.ImageData

artifact_folder = '/zfs/wficai/pda/model_run_artifacts/20220818_multitask_224x224'

with open(artifact_folder + '/config.json', 'r') as f: 
    cfg = json.load(f)

# put all config variables in scope to avoid the need to laboriously index cfg
for k, v in cfg.items():
    v = f"'{v}'" if type(v)==str else v
    exec(f"{k}={v}")
del cfg

# optionally override settings
view_filter = ['pdaView', 'pdaRelatedView', 'nonPDAView']
mode_filter = ['2d', 'color', 'color_compare']
device = torch.device('cuda:1')  # you may need 'cuda:0'

tfms = my_transforms.ImageTransforms(res)
tfms_test = tfms.get_transforms(transforms['test'])

df_test = pd.read_csv(f'{artifact_folder}/{out_paths["test"]}')
d_test = ImageData(df_test, transforms = tfms_test, mode_filter = mode_filter, view_filter = view_filter)
dl_test = DataLoader(d_test, batch_size=bs_test, num_workers=num_workers)

print("Number of frames after filtering:", len(d_test.data))

# create model
encoder = timm.create_model(model, pretrained=pretrained, num_classes=1, in_chans=3, drop_rate=dropout)
clf = MultiTaskFrameClassifier(encoder).to(device)    
clf.load_state_dict(torch.load(f"{artifact_folder}/model_checkpoint.ckpt"))
clf.eval()
loss_function = MultiTaskFrameClassifier.multi_task_loss

target_ls = []
output_ls = []
study_ls = []
video_ls = []
view_ls = []
mode_ls = []
losses = []

for ix, batch in enumerate(dl_test):
    print(f"Batch {ix+1}", end = "\r")
    inputs = batch['img'].to(device)
    targets = {k: batch[k].to(device).type(torch.float32) for k in ['trg_type', 'trg_mode', 'trg_view']}
    
    target_ls.append(batch['trg_type'].numpy())
    view_ls.append(batch['trg_view'].numpy())
    mode_ls.append(batch['trg_mode'].numpy())
    study_ls += batch['study']
    video_ls += batch['video']

    with torch.no_grad():
        outputs = clf(inputs)
        output_ls.append(outputs)
        loss = loss_function(outputs, targets, weights)
        losses.append(loss)

# Compute Metrics

def compute_metrics(y_true, y_pred, thresh=0.5):
    mets = dict()
    is_multiclass = (len(y_pred.shape)==2) & (y_pred.shape[-1]>1)
    
    if not is_multiclass:
        y_pred_cls = (y_pred>thresh).astype(int)
    else:
        y_pred_cls = np.argmax(y_pred, axis=-1)
    
    mets['num_samples'] = len(y_true)
    mets['roc_auc'] = skmet.roc_auc_score(y_true, y_pred, multi_class='ovr')
    mets['accuracy'] = skmet.accuracy_score(y_true, y_pred_cls)
    mets['sensitivity'] = skmet.recall_score(y_true, y_pred_cls, average='micro')
    
    if not is_multiclass:
        mets['specificity'] = skmet.recall_score(y_true, y_pred_cls, pos_label=0)
    
    return mets

pred_type = np.concatenate([out['type'].cpu().squeeze() for out in output_ls])
pred_type = 1/(1+np.exp(-pred_type))
trg_type = np.concatenate([trg.squeeze() for trg in target_ls])

pred_view = np.concatenate([out['view'].cpu().squeeze() for out in output_ls])
pred_view = np.exp(pred_view) / np.exp(pred_view).sum(axis=-1, keepdims=True)
trg_view = np.concatenate([trg.squeeze() for trg in view_ls])

pred_mode = np.concatenate([out['mode'].cpu().squeeze() for out in output_ls])
pred_mode = np.exp(pred_mode) / np.exp(pred_mode).sum(axis=-1, keepdims=True)
trg_mode = np.concatenate([trg.squeeze() for trg in mode_ls])

### PDA

df_pda = pd.DataFrame({'type': trg_type, 'pred': pred_type, 'video': video_ls, 'mode': trg_mode, 'view': trg_view})
df_pda_unmapped = df_pda.copy()
df_pda['mode'] = df_pda['mode'].map(ImageData.inv_mode_map)
df_pda['view'] = df_pda['view'].map(ImageData.inv_view_map)
df_pda.head()

print("frame-level scores:")
compute_metrics(df_pda['type'], df_pda['pred'])

grouped_results = df_pda.groupby(['view', 'mode']).apply(lambda x: compute_metrics(x['type'], x['pred']))
grouped_results = pd.DataFrame(grouped_results.tolist(), index=grouped_results.index)
grouped_results

print('video-level-scores')
df_pda_vid = df_pda.groupby(['type', 'video', 'mode', 'view'], as_index=False).agg('mean')
compute_metrics(df_pda_vid['type'], df_pda_vid['pred'])

grouped_results = df_pda_vid.groupby(['view', 'mode']).apply(lambda x: compute_metrics(x['type'], x['pred']))
grouped_results = pd.DataFrame(grouped_results.tolist(), index=grouped_results.index)
grouped_results

df_pda_vid_goodviews = df_pda_vid.query('mode!="2d" and view!="nonPDAView"')
compute_metrics(df_pda_vid_goodviews['type'], df_pda_vid_goodviews['pred'])

print("View prediction")
compute_metrics(trg_view, pred_view)

df_pda_vid.view

print("Mode prediction")
compute_metrics(trg_mode, pred_mode)
